move helper function to _test_commons.py to avoid circular dependency
