+++ orttraining/orttraining/test/python/_test_commons.py
@@ -4,10 +4,11 @@ import sys
 import subprocess
 import copy
 import numpy as np
+from numpy.testing import assert_allclose
 import torch
 import onnx
 
-from onnxruntime.training import optim
+from onnxruntime.training import optim, _utils
 
 def _single_run(execution_file, scenario, checkopint_dir = None):
     cmd = [sys.executable, execution_file]
@@ -136,3 +137,67 @@ def generate_dummy_optim_state(model, optimizer):
             'optimizer_name': optimizer.name
         }
     }
+
+def _load_pytorch_transformer_model(device, dynamic_axes=False, legacy_api=False):
+    # Loads external Pytorch TransformerModel into utils
+    pytorch_transformer_path = os.path.join('samples', 'python', 'pytorch_transformer')
+    pt_model_path = os.path.join(pytorch_transformer_path, 'pt_model.py')
+    pt_model = _utils.import_module_from_file(pt_model_path)
+    ort_utils_path = os.path.join(pytorch_transformer_path, 'ort_utils.py')
+    ort_utils = _utils.import_module_from_file(ort_utils_path)
+    utils_path = os.path.join(pytorch_transformer_path, 'utils.py')
+    utils = _utils.import_module_from_file(utils_path)
+
+    # Modeling
+    model = pt_model.TransformerModel(28785, 200, 2, 200, 2, 0.2).to(device)
+    my_loss = ort_utils.my_loss
+    if legacy_api:
+        if dynamic_axes:
+            model_desc = ort_utils.legacy_transformer_model_description_dynamic_axes()
+        else:
+            model_desc = ort_utils.legacy_transformer_model_description()
+    else:
+        if dynamic_axes:
+            model_desc = ort_utils.transformer_model_description_dynamic_axes()
+        else:
+            model_desc = ort_utils.transformer_model_description()
+
+
+    # Preparing data
+    train_data, val_data, test_data = utils.prepare_data(device, 20, 20)
+    return model, model_desc, my_loss, utils.get_batch, train_data, val_data, test_data
+
+def assert_all_states_close_ort(state_dict_pre_checkpoint, state_dict_post_checkpoint, reshape_states=False):
+    """Assert that the two ORTTrainer (hierarchical) state dictionaries are very close for all states"""
+
+    assert ('model' in state_dict_pre_checkpoint) == ('model' in state_dict_post_checkpoint)
+    assert ('optimizer' in state_dict_pre_checkpoint) == ('optimizer' in state_dict_post_checkpoint)
+
+    if 'model' in state_dict_pre_checkpoint:
+        for model_state_key in state_dict_pre_checkpoint['model']['full_precision']:
+            if reshape_states:
+                assert_allclose(state_dict_pre_checkpoint['model']['full_precision'][model_state_key],
+                    state_dict_post_checkpoint['model']['full_precision'][model_state_key]\
+                        .reshape(state_dict_pre_checkpoint['model']['full_precision'][model_state_key].shape))
+            else:
+                assert_allclose(state_dict_pre_checkpoint['model']['full_precision'][model_state_key],
+                    state_dict_post_checkpoint['model']['full_precision'][model_state_key])
+
+    if 'optimizer' in state_dict_pre_checkpoint:
+        for model_state_key in state_dict_pre_checkpoint['optimizer']:
+            for optimizer_state_key in state_dict_pre_checkpoint['optimizer'][model_state_key]:
+                if reshape_states:
+                    assert_allclose(state_dict_pre_checkpoint['optimizer'][model_state_key][optimizer_state_key],
+                        state_dict_post_checkpoint['optimizer'][model_state_key][optimizer_state_key]\
+                            .reshape(state_dict_pre_checkpoint['optimizer'][model_state_key][optimizer_state_key].shape))
+                else:
+                    assert_allclose(state_dict_pre_checkpoint['optimizer'][model_state_key][optimizer_state_key],
+                        state_dict_post_checkpoint['optimizer'][model_state_key][optimizer_state_key])
+
+def assert_all_states_close_pytorch(state_dict_pre_checkpoint, pytorch_model):
+    """Assert that the state_dict_pre_checkpoint state dictionary is very close to the one extracted from the pytorch model after loading"""
+
+    pytorch_model.load_state_dict(state_dict_pre_checkpoint)
+    state_dict_pytorch = pytorch_model.state_dict()
+    for key, value in state_dict_pytorch.items():
+        assert_allclose(value, state_dict_pre_checkpoint[key])
